Fixed previous commits error
program had issue of calling i2c write functions (to pass settings
from main arguments) before enabling i2c capability in initilisation.
To fix this the program was changed so the arguments are read and
saved into variables, which are then trickled down as parameters
for their respective initialisation functions.