house-cleaning. Investigating reading magnetometer values.
The magnetometer reads all zeroes, It communicates fine through i2c,
as verified by the i2c dev tools on the raspberry pi, but it seems
to only want to give zeroes when read like this. After the running
the program the i2c dev tools were used to read an axis data
register and it gave 0 as result, doing it immediately a second
time yeilded an actual result of 0xaf or whatever. and continous
reading of this lead to it giving the same actual result of 0xaf
or whatever over and over, as expected because it was not put it
 into a wake mode via the CNTL register everytime as it needs to be.